// src/constants/enums/login.ts
export enum LoginForm {
  id = "login_form",

  autocomplete_inputs = "off",
  input_size = "medium",
  button_size = "large",

  email_input_id = "email_input_id",
  email_input_name = "email_input",
  email_input_label = "",
  email_input_role = "email-input",
  email_input_arial_label = "email-input",
  email_input_placeholder = "Your email address",
  email_input_type = "email",

  password_input_id = "password_input_id",
  password_input_name = "password_input",
  password_input_label = "",
  password_input_role = "password-input",
  password_input_arial_label = "password-input",
  password_input_placeholder = "Your password",
  password_input_type_hidden = "password",
  password_input_type_show = "text",
}

export enum LoginFormTexts {
  unregistered_user = "Don't have an account?",
  form_title = "Login",
  register_button = "Join now",
  login_button = "Login",
  forgot_password = "Forgot password?",
  forgot_password_button = "Reset password"
}